+++ lib/hcp.rb
@@ -35,6 +35,7 @@ require 'hcp/concerns/scheduled'
 
 require 'hcp/resources/address'
 require 'hcp/resources/booking_window'
+require 'hcp/resources/company'
 require 'hcp/resources/employee'
 require 'hcp/resources/line_item'
 require 'hcp/resources/note'